5o caso test superado

+++ src/main/java/org/mvpigs/cotxox/repo/CarreraRepo.java
@@ -3,5 +3,5 @@ package org.mvpigs.cotxox.repo;
 import org.mvpigs.cotxox.domain.Carrera;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarreraRepo extends CrudRepository<Carrera, Integer> {
+public interface CarreraRepo extends CrudRepository<Carrera, Long> {
 }
